refactor(daily): complete TimeFlow UI3層分離 (Phase 1-3)
- Phase 1: Extract types, constants, utilities (1877→1600 lines)
- Phase 2: Extract 9 view components to views/ (1600→600 lines)
- Phase 3: Extract useTimeFlowState hook + thin page (600→3 lines)
- Add 3 safety-net tests for hook state transitions
- Pages entry point is now a re-export only

1877 → 3 lines: architecture maturity, not just cleanup
