+++ src/features/daily/components/time-flow/index.ts
@@ -1,0 +1,6 @@
+// ---------------------------------------------------------------------------
+// Barrel export for time-flow feature module
+// ---------------------------------------------------------------------------
+export { useTimeFlowState } from './hooks/useTimeFlowState';
+export type { UseTimeFlowStateResult } from './hooks/useTimeFlowState';
+export { default } from './TimeFlowPage';
